Fixing undefined behavior in basisu_etc.cpp
The value inside the static cast can be negative. Casting a negative floating point value to an unsigned integer is an undefined behavior and we were running into issues when executing the code with an address sanitizer turned on.

Not sure if this is the proper way to fix this but it seems to work ok. Either way, the undefined behavior should be fixed.