Make repository compliant with REUSE (https://reuse.software).
This PR simply wildcards files in the repository in the metainformation
.reuse/dep5 to express copyright and licensing information for files
that can't easily be parsed by REUSE. If any of the assertions made in
.reuse/dep5 are incorrect, we should fix that before this PR
is accepted.

This approach works OK, though a more intrusive, but superior version of
this PR would simplify copyright / license statements as follows, and
make sure they're present in every file:

// Copyright 2019-2020 Binomial LLC
// SPDX-License-Identifier: Apache-2.0

(or whatever the appropriate copyright owner and license are for each
file). If that were done, the wildcarding in .reuse/dep5 would not be
required to pass REUSE.